# 8. Sistemas distribuidos/new_app/docs/specs/swagger.yaml
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "new-app-01",
    "description" : "new-app-01",
    "contact" : {
      "name" : "mramosgarcia",
      "email" : "miguel.ramosgarcia@mercadolibre.com.co"
    },
    "version" : "last"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Scope local"
  } ],
  "paths" : {
...
    "/findSport/{name}" : {
      "get" : {
        "tags" : [ "deportista-controller" ],
        "operationId" : "findSportByName",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DificultadDeporteDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "DificultadDeporteDTO" : {
        "type" : "object",
        "properties" : {
          "nivel" : {
            "type" : "string"
          }
        }
      }
    }
  }
}