# 8. Sistemas distribuidos/new_app/src/main/java/com/mercadolibre/new_app_01/dtos/DificultadDeporteDTO.java
package com.mercadolibre.new_app_01.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class DificultadDeporteDTO {
    private String nivel;
}